logic/requirements.txt
fastapi
uvicorn
pandas
langgraph
langchain-openai
python-dotenv
prometheus-client
pytest
httpx
torch
google-genai
streamlit
plotly